fix: refactor to make more modular
